+++ scurrypy/models/__init__.py
@@ -3,6 +3,13 @@
 from .application import ApplicationFlags, ApplicationModel
 from .attachment import AttachmentModel
 from .channel import ChannelModel, PinnedMessageModel
+from .command import (
+    ApplicationCommandTypes,
+    ApplicationCommandOptionTypes,
+    ApplicationCommandOptionChoiceModel,
+    ApplicationCommandOptionModel,
+    ApplicationCommandModel
+)
 from .emoji import EmojiModel
 from .guild_member import GuildMemberModel
 from .guild import ReadyGuildModel, GuildModel
@@ -23,6 +30,8 @@ __all__ = [
     "ApplicationFlags", "ApplicationModel",
     "AttachmentModel",
     "ChannelModel", "PinnedMessageModel",
+    "ApplicationCommandTypes", "ApplicationCommandOptionTypes", "ApplicationCommandOptionChoiceModel", 
+    "ApplicationCommandOptionModel", "ApplicationCommandModel",
     "EmojiModel",
     "GuildMemberModel",
     "ReadyGuildModel", "GuildModel",